pyproject.toml
[project]
name = "fraud-detection-xgboost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "cupy-cuda12x>=13.3.0",
    "hyperopt>=0.2.7",
    "numpy==2.0.2",
    "onnxmltools>=1.13.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "setuptools>=75.8.1",
    "xgboost>=2.1.4",
]
